Reorganized mesh classes
