+++ bundles/pixi.js/package.json
@@ -49,6 +49,7 @@
     "@pixi/loaders": "^5.0.0-alpha.2",
     "@pixi/math": "^5.0.0-alpha.2",
     "@pixi/mesh": "^5.0.0-alpha.2",
+    "@pixi/mesh-extras": "^5.0.0-alpha.2",
     "@pixi/mixin-app-loader": "^5.0.0-alpha.2",
     "@pixi/mixin-cache-as-bitmap": "^5.0.0-alpha.2",
     "@pixi/mixin-get-child-by-name": "^5.0.0-alpha.2",
