+++ packages/canvas/canvas-mesh/src/index.js
@@ -1,6 +1,6 @@
 export { default as CanvasMeshRenderer } from './CanvasMeshRenderer';
 
 import './settings';
-import './Mesh';
+import './Mesh2d';
 import './NineSlicePlane';
-import './RawMesh';
+import './Mesh';
