+++ packages/canvas/canvas-mesh/src/settings.js
@@ -3,7 +3,7 @@ import { settings } from '@pixi/settings';
 /**
  * Default `canvasPadding` for canvas-based Mesh rendering.
  *
- * @see PIXI.Mesh#canvasPadding
+ * @see PIXI.Mesh2d#canvasPadding
  * @static
  * @constant
  * @memberof PIXI.settings
