+++ packages/mesh-extras/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@pixi/mesh",
+  "name": "@pixi/mesh-extras",
   "version": "5.0.0-alpha.2",
-  "main": "lib/mesh.js",
-  "module": "lib/mesh.es.js",
-  "description": "Support for legacy browser JavaScript environments",
+  "main": "lib/mesh-extras.js",
+  "module": "lib/mesh-extras.es.js",
+  "description": "Custom Mesh display objects, like Rope and Plane",
   "author": "Mat Groves",
   "contributors": [
     "Matt Karl <matt@mattkarl.com>"
@@ -18,16 +18,21 @@
   "publishConfig": {
     "access": "public"
   },
+  "scripts": {
+    "test": "floss --path test"
+  },
   "files": [
     "lib"
   ],
   "dependencies": {
     "@pixi/constants": "^5.0.0-alpha.2",
     "@pixi/core": "^5.0.0-alpha.2",
-    "@pixi/display": "^5.0.0-alpha.2",
-    "@pixi/math": "^5.0.0-alpha.2"
+    "@pixi/math": "^5.0.0-alpha.2",
+    "@pixi/mesh": "^5.0.0-alpha.2",
+    "@pixi/utils": "^5.0.0-alpha.2"
   },
   "devDependencies": {
+    "@pixi/loaders": "^5.0.0-alpha.2",
     "floss": "^2.1.3"
   }
 }
