+++ packages/mesh-extras/src/NineSlicePlane.js
