+++ packages/mesh-extras/src/index.js
@@ -1,0 +1,4 @@
+export { default as Mesh2d } from './Mesh2d';
+export { default as Plane } from './Plane';
+export { default as NineSlicePlane } from './NineSlicePlane';
+export { default as Rope } from './Rope';
