+++ packages/mesh-extras/src/mesh.vert
