+++ packages/mesh-extras/test/resources/bitmap-1.png
