+++ packages/mesh/src/index.js
@@ -1,6 +1,2 @@
 export { default as Mesh } from './Mesh';
-export { default as RawMesh } from './RawMesh';
 export { default as MeshRenderer } from './MeshRenderer';
-export { default as Plane } from './Plane';
-export { default as NineSlicePlane } from './NineSlicePlane';
-export { default as Rope } from './Rope';
